option classes for european, american, bermudan
